# openapi.yaml
openapi: 3.1.0
info:
  title: Saerbridge Central API
  version: 1.0.0
  description: |
    Cookie-authenticated API for saerbridge.com. Production host is https://api.saerbridge.com.
    Send credentials. State-changing methods require header X-CSRF-Token from GET /api/v1/auth/session.
  contact:
    email: we@saerbridge.com
servers:
  - url: https://api.saerbridge.com
  - url: http://localhost:8080
paths:
...
  /api/v1/products/{slug}:
    get:
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200': { description: Product }
        '404': { description: Not found }
...
  /api/v1/admin/products:
    get:
      responses:
        '200': { description: Admin list }
    post:
      responses:
        '201': { description: Created }
        '403': { description: Not admin }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Secure-saerbridge.sid
...